Fix #21 key duplication issue by making leaf bindings use key release events
This change fixes the issue where the last character of key bindings was
being passed through to the opened window/action. Now we ensure that:
- Group bindings use key press events (on press)
- Leaf/terminal bindings use key release events (on release)

This way, the key press is properly consumed by the action before it
can be passed to the newly opened window.

By treating group bindings as on press, we ensure that perfomance gains
mentioned in #3 are retained